- Create a database directory under services to hold database related scripts. - Moved all existing database scripts into the new directory. - Pointed all files using the eligibility-client.js file to its new location.
