+++ alpha/internal-web/app/routes/claim-details.js
@@ -1,5 +1,5 @@
 var router = require('../routes')
-var client = require('../services/db-client')
+var client = require('../services/database/db-client')
 
 const claimsCollection = 'claims'
