+++ alpha/internal-web/app/routes/claims.js
@@ -1,5 +1,5 @@
 var router = require('../routes')
-var client = require('../services/db-client')
+var client = require('../services/database/db-client')
 var moment = require('moment')
 
 const claimsCollection = 'claims'
