+++ alpha/internal-web/app/routes/index.js
@@ -1,5 +1,5 @@
 var router = require('../routes')
-var client = require('../services/db-client')
+var client = require('../services/database/db-client')
 
 router.get('/', function (request, response, next) {
   response.render('index')
